requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
airflow-multi-dagrun==2.2.0
alembic==1.8.0
anyio==3.6.1
apache-airflow==2.3.3
apache-airflow-providers-amazon==4.0.0
apache-airflow-providers-apache-spark==3.0.0
apache-airflow-providers-cncf-kubernetes==4.1.0
apache-airflow-providers-ftp==3.0.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-mysql==3.0.0
apache-airflow-providers-slack==5.0.0
apache-airflow-providers-discord==3.0.0
apache-airflow-providers-sqlite==3.0.0
apispec==3.3.2
argcomplete==2.0.0
asn1crypto==1.5.1
astroid==2.11.3
async-generator==1.10
async-timeout==4.0.2
attrs==20.3.0
awswrangler==2.12.1
Babel==2.10.3
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
boto3==1.24.23
boto3-stubs==1.21.25
boto3-type-annotations==0.3.1
botocore==1.27.23
botocore-stubs==1.24.25
cachelib==0.9.0
cachetools==4.2.2
cattrs==1.10.0
certifi==2020.12.5
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
colorama==0.4.5
colorlog==4.8.0
commonmark==0.9.1
connexion==2.14.0
coverage==6.1.1
cron-descriptor==1.2.30
croniter==1.3.5
cryptography==36.0.2
curlify==2.2.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
discord.py==2.0.1
dnspython==2.2.1
docker==5.0.3
docutils==0.19
ecdsa==0.14.1
email-validator==1.2.1
et-xmlfile==1.1.0
Flask==2.1.2
Flask-AppBuilder==4.1.2
Flask-Babel==2.0.0
Flask-Caching==2.0.0
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
frozenlist==1.2.0
google-auth==2.9.0
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
inflection==0.5.1
iniconfig==1.1.1
iso8601==0.1.16
isodate==0.6.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
jsonpath-ng==1.5.3
jsonschema==4.6.1
kubernetes==23.6.0
lazy-object-proxy==1.7.1
linkify-it-py==2.0.0
lockfile==0.12.2
lxml==4.9.1
Mako==1.2.1
Markdown==3.3.7
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mccabe==0.7.0
mdit-py-plugins==0.3.0
mdurl==0.1.1
more-itertools==8.10.0
moto==3.1.16
multidict==5.2.0
mypy==0.910
mypy-boto3-cognito-idp==1.21.23.post1
mypy-boto3-rds==1.24.23
mypy-boto3-redshift-data==1.24.11.post3
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
numpy==1.22.4
oauthlib==3.2.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
openpyxl==3.0.9
opensearch-py==1.1.0
oracledb==1.0.2
outcome==1.1.0
packaging==21.3
pandas==1.4.3
paramiko==2.11.0
pathspec==0.9.0
pdpyras==4.5.0
pendulum==2.1.2
pg8000==1.21.3
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
prison==0.2.1
progressbar2==3.55.0
protobuf==3.20.0
psutil==5.9.1
psycopg2-binary==2.9.1
py==1.10.0
py4j==0.10.9.5
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.12.0
PyJWT==2.4.0
pylint==2.13.7
pylint-airflow==0.1.0a1
PyMySQL==1.0.2
PyNaCl==1.5.0
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyspark==3.3.0
pytest==6.2.5
pytest-base-url==1.4.2
pytest-cov==3.0.0
pytest-datadir==1.3.1
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-mock==3.6.1
pytest-mock-resources==2.2.5
pytest-variables==1.9.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-editor==1.0.4
python-jose==3.3.0
python-nvd3==0.15.0
python-slugify==6.1.2
python-utils==3.1.0
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
redshift-connector==2.0.908
requests==2.28.0
requests-aws4auth==1.1.2
requests-mock==1.9.3
requests-oauthlib==1.3.1
responses==0.15.0
rfc3986==1.5.0
rich==12.4.4
rsa==4.8
s3transfer==0.6.0
scramp==1.4.1
sentry-sdk==1.4.3
setproctitle==1.2.3
six==1.16.0
slack-sdk==3.17.2
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.27
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.9
SQLAlchemy-Utils==0.38.2
sshtunnel==0.4.0
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
trio==0.19.0
trio-websocket==0.9.2
types-requests==2.26.3
typing_extensions==4.3.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.9
watchtower==2.0.1
websocket-client==1.3.3
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.0.0
WTForms==2.3.3
xlrd==2.0.1
xmltodict==0.12.0
yarl==1.7.2
zipp==3.8.0
